Adjust command ids move to 118-124 after main's folder_up/prev/next; the
adjust pane now uses main's pane focus (focus_kind::pane) instead of text
focus; exports pass main's carried metadata through the colour bake; the
HLSL camera cbuffer carries main's inverse map and PR 11's uniforms.

+++ src/edit/encode.cpp
@@ -179,7 +179,8 @@ result<std::vector<std::uint8_t>> encode_png(const codec::raster& img, const met
   ihdr.width = img.width;
   ihdr.height = img.height;
   ihdr.bit_depth = 8;
-  ihdr.color_type = opaque ? SPNG_COLOR_TYPE_TRUECOLOR : SPNG_COLOR_TYPE_TRUECOLOR_ALPHA;
+  ihdr.color_type = static_cast<std::uint8_t>(opaque ? SPNG_COLOR_TYPE_TRUECOLOR
+                                                   : SPNG_COLOR_TYPE_TRUECOLOR_ALPHA);
   if (spng_set_ihdr(g.ctx, &ihdr) != 0) return err(status::invalid_arg);
 
   // libspng copies nothing: the buffers below must outlive the encode.
